refactor: split the CameraGstreamer classes in two different files
